requirements.txt

flask~=2.2.2
python-dotenv~=0.21.0
feedparser~=6.0.8